edit tests

+++ app.py
@@ -312,7 +312,7 @@ def messages_add():
 def messages_show(message_id):
     """Show a message."""
 
-    msg = Message.query.get(message_id)
+    msg = Message.query.get_or_404(message_id)
     return render_template('messages/show.html', message=msg)
 
 
@@ -320,13 +320,16 @@ def messages_show(message_id):
 @check_g_user
 def messages_destroy(message_id):
     """Delete a message."""
+    
+    msg = Message.query.get_or_404(message_id)
+    if (g.user.id == msg.user_id):
+        db.session.delete(msg)
+        db.session.commit()
 
-    msg = Message.query.get(message_id)
-    db.session.delete(msg)
-    db.session.commit()
-
-    flash ('Deleted message!', 'info')
-    return redirect (url_for("users_show", user_id=g.user.id))
+        flash ('Deleted message!', 'info')
+        return redirect (url_for("users_show", user_id=g.user.id))
+    flash("Access unauthorized.", "danger")
+    return redirect("/")
 
 
 ##############################################################################
